+++ pocket-services/src/result-set.cpp
@@ -32,7 +32,7 @@ result_set::result_set(class database& database, const std::string& query, const
     sqlite3_stmt *stmt = nullptr;
 
     debug(typeid(*this).name(), query);
-    statement_stat = sqlite3_prepare_v3(database.db, query.c_str(), query.length(), 0, &stmt, nullptr);
+    statement_stat = sqlite3_prepare_v3(database.db, query.c_str(), static_cast<int>(query.length()), 0, &stmt, nullptr);
     if(statement_stat == SQLITE_OK )
     {
         for(int i = 1; auto &&param : parameters) {
